corrected a variable type for python "b", i.e., from uint16 to uint8. this fixes the recent build failure on sparc: http://bind10.isc.org/~tester/builder//BIND10/20100929022000-Solaris10-sparc/logs/unittests.out
directly committing to trunk, as I believe it's quite trivial and it's a
sort of urgent care fix.


git-svn-id: svn://bind10.isc.org/svn/bind10/trunk@3060 e5f2f494-b856-4b98-b285-d166d9295462

+++ src/lib/dns/python/opcode_python.cc
@@ -159,7 +159,7 @@ PyTypeObject opcode_type = {
 
 int
 Opcode_init(s_Opcode* const self, PyObject* args) {
-    uint16_t code = 0;
+    uint8_t code = 0;
     if (PyArg_ParseTuple(args, "b", &code)) {
         try {
             self->opcode = new Opcode(code);
